[Dev] Generic receiver, part 8.
