+++ include/cocaine/framework/receiver.hpp
@@ -14,10 +14,6 @@
 #include <boost/variant.hpp>
 
 #include <cocaine/common.hpp>
-#include <cocaine/idl/primitive.hpp>
-#include <cocaine/idl/streaming.hpp>
-#include <cocaine/rpc/asio/decoder.hpp>
-#include <cocaine/utility.hpp>
 
 #include "cocaine/framework/error.hpp"
 #include "cocaine/framework/forwards.hpp"
@@ -28,6 +24,8 @@ namespace cocaine {
 
 namespace framework {
 
+/// The basic receiver provides an interface to extract incoming MessagePack'ed payloads from the
+/// session.
 template<class Session>
 class basic_receiver_t {
     std::uint64_t id;
@@ -39,182 +37,10 @@ public:
 
     ~basic_receiver_t();
 
+    /// Returns a future with a decoded message received from the session.
+    ///
+    /// This future may throw std::system_error on any network failure.
     auto recv() -> task<decoded_message>::future_type;
-};
-
-/// Unpacks message pack object into concrete types.
-/// \internal
-template<class Receiver>
-class slot_unpacker {
-public:
-    typedef typename detail::result_of<Receiver>::type variant_type;
-    typedef std::function<variant_type(std::shared_ptr<basic_receiver_t<typename Receiver::session_type>>, const msgpack::object&)> function_type;
-
-public:
-    slot_unpacker(std::vector<function_type>& unpackers) :
-        unpackers(unpackers)
-    {}
-
-    template<class U>
-    void operator()(U*) {
-        unpackers.emplace_back(detail::unpacker<U, typename Receiver::session_type>());
-    }
-
-private:
-    std::vector<function_type>& unpackers;
-};
-
-template<class T, class Session>
-class slot_unpacker<receiver<io::streaming_tag<T>, Session>> {
-public:
-    typedef typename detail::variant_of<io::streaming_tag<T>>::type variant_type;
-
-    typedef std::function<variant_type(const msgpack::object&)> function_type;
-
-public:
-    slot_unpacker(std::vector<function_type>& unpackers) :
-        unpackers(unpackers)
-    {}
-
-    template<class U>
-    void operator()(U*) {
-        unpackers.emplace_back(detail::unpacker<U, Session>());
-    }
-
-private:
-    std::vector<function_type>& unpackers;
-};
-
-template<class Receiver>
-static
-std::vector<typename slot_unpacker<Receiver>::function_type>
-make_slot_unpacker() {
-    typedef slot_unpacker<Receiver> unpacker_type;
-
-    std::vector<typename unpacker_type::function_type> result;
-
-    boost::mpl::for_each<
-        typename boost::mpl::transform<
-            typename unpacker_type::variant_type::types,
-            std::add_pointer<boost::mpl::_1>
-        >::type
-    >(unpacker_type(result));
-
-    return result;
-}
-
-template<class T>
-struct unpacked_result;
-
-template<class T>
-struct unpacked_result<std::tuple<T>> {
-    typedef T type;
-
-    static inline
-    T unpack(std::tuple<T>& from) {
-        return std::get<0>(from);
-    }
-};
-
-template<class... Args>
-struct unpacked_result<std::tuple<Args...>> {
-    typedef std::tuple<Args...> type;
-
-    static inline
-    std::tuple<Args...> unpack(std::tuple<Args...>& from) {
-        return from;
-    }
-};
-
-/// Helper trait, that simplifies event receiving, that use one of the common protocols.
-template<class T, class Session>
-struct from_receiver {
-private:
-    typedef typename detail::result_of<receiver<T, Session>>::type variant_type;
-
-public:
-    typedef variant_type result_type;
-
-    static inline
-    result_type
-    transform(variant_type& value) {
-        return value;
-    }
-};
-
-template<class T, class Session>
-struct from_receiver<io::primitive_tag<T>, Session> {
-private:
-    typedef typename detail::result_of<receiver<io::primitive_tag<T>, Session>>::type variant_type;
-
-    typedef typename boost::mpl::at<typename variant_type::types, boost::mpl::int_<0>>::type value_type;
-    typedef typename boost::mpl::at<typename variant_type::types, boost::mpl::int_<1>>::type error_type;
-
-public:
-    typedef typename unpacked_result<value_type>::type result_type;
-
-    static
-    result_type
-    transform(variant_type& value) {
-        return boost::apply_visitor(visitor_t(), value);
-    }
-
-private:
-    struct visitor_t : public boost::static_visitor<result_type> {
-        result_type operator()(value_type& value) const {
-            return unpacked_result<value_type>::unpack(value);
-        }
-
-        result_type operator()(error_type& error) const {
-            throw cocaine::framework::response_error(error);
-        }
-    };
-};
-
-template<class T, class Session>
-struct from_receiver<io::streaming_tag<T>, Session> {
-private:
-    typedef typename detail::variant_of<io::streaming_tag<T>>::type variant_type;
-
-    typedef typename boost::mpl::at<typename variant_type::types, boost::mpl::int_<0>>::type value_type;
-    typedef typename boost::mpl::at<typename variant_type::types, boost::mpl::int_<1>>::type error_type;
-    typedef typename boost::mpl::at<typename variant_type::types, boost::mpl::int_<2>>::type choke_type;
-
-public:
-    typedef boost::optional<typename unpacked_result<value_type>::type> result_type;
-
-    static
-    result_type
-    transform(variant_type& value) {
-        return boost::apply_visitor(visitor_t(), value);
-    }
-
-private:
-    struct visitor_t : public boost::static_visitor<result_type> {
-        result_type operator()(value_type& value) const {
-            return boost::make_optional(unpacked_result<value_type>::unpack(value));
-        }
-
-        result_type operator()(error_type& error) const {
-            throw cocaine::framework::response_error(error);
-        }
-
-        result_type operator()(choke_type&) const {
-            return boost::none;
-        }
-    };
-};
-
-template<class Session, class Result>
-struct unpacker_factory {
-    typedef Result result_type;
-
-    template<class T>
-    static inline
-    result_type
-    apply() {
-        return detail::unpacker<T, Session>();
-    }
 };
 
 template<class T, class Session>
@@ -319,14 +145,14 @@ const std::array<
     typename receiver<T, Session>::unpacker_type,
     boost::mpl::size<typename receiver<T, Session>::result_type::types>::value
 > receiver<T, Session>::unpackers =
-    meta::to_array<typename receiver<T, Session>::result_type::types, unpacker_factory<Session, typename receiver<T, Session>::unpacker_type>>::make();
+    meta::to_array<typename receiver<T, Session>::result_type::types, detail::unpacker_factory<Session, typename receiver<T, Session>::unpacker_type>>::make();
 
 template<class T, class Session>
 const std::array<
     typename receiver<io::streaming_tag<T>, Session>::unpacker_type,
     boost::mpl::size<typename receiver<io::streaming_tag<T>, Session>::result_type::types>::value
 > receiver<io::streaming_tag<T>, Session>::unpackers =
-    meta::to_array<typename receiver<io::streaming_tag<T>, Session>::result_type::types, unpacker_factory<Session, typename receiver<io::streaming_tag<T>, Session>::unpacker_type>>::make();
+    meta::to_array<typename receiver<io::streaming_tag<T>, Session>::result_type::types, detail::unpacker_factory<Session, typename receiver<io::streaming_tag<T>, Session>::unpacker_type>>::make();
 
 } // namespace framework
